typescript: Convert Tier model
